do not link jscip to ${JNI_LIBRARIES}
CMakeLists.txt: It is not necessary to link jscip to ${JNI_LIBRARIES},
i.e., libjvm.so and libjawt.so. The code SWIG generates does not
actually use any symbol from those libraries. They just ended up
unnecessarily linked, which is a problem especially for libjawt.so that
is not available in headless OpenJDK JRE installations from split
GNU/Linux distribution packages.