CMakeLists.txt: Strictly honor user-specified SCIP_DIR.
If the user has explicitly passed a SCIP_DIR, only look there and not in
the default search path. Otherwise, we may end up silently using a
different, binary-incompatible SCIP from another directory that happens
to be in the default search path, instead of the one the user intended
(because the directory was not correctly specified). This is
particularly an issue on Windows, but can happen on all operating
systems when working with SCIP directly from a build directory, without
installing it to the default prefix.

We have encountered this on a Windows system with both SCIP 7.0.3 and
8.0.0 installed, when specifying a wrong subdirectory for the search
path made it not find the SCIP 8.0.0 config, and silently pick up 7.0.3
from the search path instead.

The versions are not binary-compatible due to the macros implementing at
least SCIPinfinity, which end up poking directly in structures whose
layout changed between the versions instead of doing a function call
over the function ABI that has not changed.