CMakeLists.txt: Do not require JNI AWT library. (#41)
The list of components to find in FindJNI.cmake defaults to "JVM;AWT".
But the AWT library is not present in headless installations. So specify
only JVM as the only required component. (Use the syntax without
"COMPONENTS" for backwards compatibility with CMake 2.8, which only
allows EITHER "REQUIRED" OR "COMPONENTS", not both.)

Technically, the JVM library is not required either, but we cannot pass
an empty list here, because FindJNI.cmake then thinks we passed no list
at all and just uses the default. Passing a dummy component name does
not work either, because it then complains that the dummy component was
not found. So this is the best we can do for now, and better than the
previous status quo.

Fixes #41.