WIP: add CustomExpressionHandler example
examples/CustomExpressionHandler.java: New example. Example how to
create a problem with a custom expression handler.
Implements: https://www.minlplib.org/prob10.html using a custom square
operator.

This is WIP because the interval arithmetic does not yet use proper
rounding. Doing so would require either wrapping the SCIP interval
arithmetic functions in JSCIPOpt, which is not done yet, or dragging in
a Java interval arithmetic implementation, which I obviously do not want
to do for this simple example.